+++ src/main/kotlin/racingcar/domain/movingstrategy/MovingStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar.movingstrategy
+package racingcar.domain.movingstrategy
 
 interface MovingStrategy {
     fun isMovable(): Boolean
