+++ src/main/kotlin/racingcar/domain/numbergenerator/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar.numbergenerator
+package racingcar.domain.numbergenerator
 
 interface NumberGenerator {
     fun get(): Int
